+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file.
 ## [v1.1.0 hot-fix-2] - 2025-02-??
 
 - Fixed Egg Carrier transformed when starting from the inside.
+- Fixed extended ring capacity not working for Big.
 
 ## [v1.1.0 hot-fix] - 2025-02-12